Move Context to a separate file #398
